requirements.txt
pandas
matplotlib
fpdf
flask
colorama
python-dotenv

